+++ crates/subcoin-script/src/tests.rs
@@ -1,0 +1,1 @@
+pub mod interpreter;
